Perf: Extract Search and Workspaces tab buttons to isolate Onyx subscriptions
